Move all resources related scripts to stanza/resource
