+++ stanza/resource/prepare_resources.py
