+++ stanza/resource/resources.py
